// src/schedule/dto/google.dto.ts
export class SyncGoogleResult {
  syncGoogleUserToken?: string;
  syncGoogleUserAccessToken?: string;
  syncGoogleUserCalendarId?: string;
  syncGoogleUserLastModified?: string;
  syncGoogleUserGoogleLastModified?: string;
  userId?: string;
  groupId?: string;
  resourceId?: string;
  timezone?: string;
}